+++ mycodo/functions/display_generic_lcd_20x4_i2c.py
@@ -1,6 +1,6 @@
 # coding=utf-8
 #
-#  lcd_generic_16x2_i2c.py - Function to output to LCD
+#  lcd_generic_20x4_i2c.py - Function to output to LCD
 #
 #  Copyright (C) 2015-2020 Kyle T. Gabriel <mycodo@kylegabriel.com>
 #
@@ -44,8 +44,8 @@ from mycodo.utils.lcd import format_measurement_line
 from mycodo.utils.system_pi import cmd_output
 
 # Set to how many lines the LCD has
-lcd_lines = 2
-lcd_x_characters = 16
+lcd_lines = 4
+lcd_x_characters = 20
 
 
 def execute_at_creation(error, new_func, dict_functions=None):
@@ -178,13 +178,13 @@ def constraints_pass_positive_value(mod_controller, value):
 
 
 FUNCTION_INFORMATION = {
-    'function_name_unique': 'display_generic_lcd_16x2_i2c',
-    'function_name': 'Display: Generic LCD 16x2 (I2C)',
+    'function_name_unique': 'display_generic_lcd_20x4_i2c',
+    'function_name': 'Display: Generic LCD 20x4 (I2C)',
     'function_library': 'smbus2',
     'execute_at_creation': execute_at_creation,
     'execute_at_modification': execute_at_modification,
 
-    'message': 'This Function outputs to a generic 16x2 LCD display via I2C. Since this display can show 2 lines at a time, channels are added in sets of 2 when Number of Line Sets is modified. Every Period, the LCD will refresh and display the next set of lines. Therefore, the first 2 lines that are displayed are channels 0 and 1, then 2 and 3, and so on. After all channels have been displayed, it will cycle back to the beginning.',
+    'message': 'This Function outputs to a generic 20x4 LCD display via I2C. Since this display can show 4 lines at a time, channels are added in sets of 4 when Number of Line Sets is modified. Every Period, the LCD will refresh and display the next set of lines. Therefore, the first 4 lines that are displayed are channels 0, 1, 2, and 3, then 4, 5, 6, and 7, and so on. After all channels have been displayed, it will cycle back to the beginning.',
 
     'custom_options': [
         {
@@ -297,7 +297,7 @@ class CustomModule(AbstractFunction):
         self.timer_loop = time.time()
         self.line_sets = []
         self.current_line_set = 0
-        self.line_y_dimensions = [0, 8]
+        self.line_y_dimensions = [0, 8, 16, 24]
         self.pad = -2
 
         # Initialize custom options
@@ -436,8 +436,7 @@ class CustomModule(AbstractFunction):
 
         # Display lines
         self.lcd.lcd_init()
-        self.lcd.lcd_write_lines(
-            lines_display[0], lines_display[1], lines_display[2],lines_display[3])
+        self.lcd.lcd_write_lines(lines_display[0], lines_display[1], "", "")
 
     def stop_function(self):
         self.lcd.lcd_init()