+++ mycodo/functions/display_ssd1309_oled_128x64_i2c_03.py
@@ -1,6 +1,6 @@
 # coding=utf-8
 #
-#  lcd_generic_16x2_i2c.py - Function to output to LCD
+#  lcd_ssd1309_i2c.py - Function to output to LCD
 #
 #  Copyright (C) 2015-2020 Kyle T. Gabriel <mycodo@kylegabriel.com>
 #
@@ -44,8 +44,8 @@ from mycodo.utils.lcd import format_measurement_line
 from mycodo.utils.system_pi import cmd_output
 
 # Set to how many lines the LCD has
-lcd_lines = 2
-lcd_x_characters = 16
+lcd_lines = 8
+lcd_x_characters = 21
 
 
 def execute_at_creation(error, new_func, dict_functions=None):
@@ -178,13 +178,29 @@ def constraints_pass_positive_value(mod_controller, value):
 
 
 FUNCTION_INFORMATION = {
-    'function_name_unique': 'display_generic_lcd_16x2_i2c',
-    'function_name': 'Display: Generic LCD 16x2 (I2C)',
-    'function_library': 'smbus2',
+    'function_name_unique': 'display_ssd1309_oled_128x64_i2c_03',
+    'function_name': 'Display: SSD1309 OLED 128x64 (I2C) Test 03',
+    'function_library': 'luma.oled',
     'execute_at_creation': execute_at_creation,
     'execute_at_modification': execute_at_modification,
 
-    'message': 'This Function outputs to a generic 16x2 LCD display via I2C. Since this display can show 2 lines at a time, channels are added in sets of 2 when Number of Line Sets is modified. Every Period, the LCD will refresh and display the next set of lines. Therefore, the first 2 lines that are displayed are channels 0 and 1, then 2 and 3, and so on. After all channels have been displayed, it will cycle back to the beginning.',
+    'message': 'This Function outputs to a 128x64 SSD1309 OLED display via I2C. Since this display can show 8 lines at a time, channels are added in sets of 8 when Number of Line Sets is modified. Every Period, the LCD will refresh and display the next 8 lines. Therefore, the first 8 lines that are displayed are channels 0 - 7, then 8 - 15, and so on. After all channels have been displayed, it will cycle back to the beginning.',
+
+    'options_enabled': [
+        'custom_options'
+    ],
+
+    'dependencies_module': [
+        ('pip-pypi', 'usb.core', 'pyusb==1.1.1'),
+        ('pip-pypi', 'luma.oled', 'luma.oled==3.8.1'),
+        ('pip-pypi', 'PIL', 'Pillow==8.1.2'),
+        ('apt', 'libjpeg-dev', 'libjpeg-dev'),
+        ('apt', 'zlib1g-dev', 'zlib1g-dev'),
+        ('apt', 'libfreetype6-dev', 'libfreetype6-dev'),
+        ('apt', 'liblcms2-dev', 'liblcms2-dev'),
+        ('apt', 'libopenjp2-7', 'libopenjp2-7'),
+        ('apt', 'libtiff5', 'libtiff5')
+    ],
 
     'custom_options': [
         {
@@ -199,7 +215,7 @@ FUNCTION_INFORMATION = {
         {
             'id': 'i2c_address',
             'type': 'text',
-            'default_value': '0x20',
+            'default_value': '0x3c',
             'required': True,
             'name': TRANSLATIONS['i2c_location']['title'],
             'phrase': TRANSLATIONS['i2c_location']['phrase']
@@ -220,7 +236,15 @@ FUNCTION_INFORMATION = {
             'constraints_pass': constraints_pass_positive_value,
             'name': 'Number of Line Sets',
             'phrase': 'How many sets of lines to cycle on the LCD'
-        }
+        },
+        {
+            'id': 'pin_reset',
+            'type': 'integer',
+            'default_value': 17,
+            'required': True,
+            'name': 'Reset Pin',
+            'phrase': 'The pin (BCM numbering) connected to RST of the display'
+        },
     ],
 
     'custom_channel_options': [
@@ -264,15 +288,6 @@ FUNCTION_INFORMATION = {
             'phrase': 'The maximum allowed age of the measurement'
         },
         {
-            'id': 'measure_decimal',
-            'type': 'integer',
-            'default_value': 1,
-            'required': True,
-            'constraints_pass': constraints_pass_positive_value,
-            'name': 'Measurement Decimal',
-            'phrase': 'The number of digits after the decimal'
-        },
-        {
             'id': 'text',
             'type': 'text',
             'default_value': 'Text',
@@ -292,12 +307,12 @@ class CustomModule(AbstractFunction):
         super(CustomModule, self).__init__(function, testing=testing, name=__name__)
 
         self.options_channels = {}
-        self.lcd = None
+        self.device = None
         self.canvas = None
         self.timer_loop = time.time()
         self.line_sets = []
         self.current_line_set = 0
-        self.line_y_dimensions = [0, 8]
+        self.line_y_dimensions = [0, 8, 16, 24, 32, 40, 48, 56]
         self.pad = -2
 
         # Initialize custom options
@@ -305,6 +320,7 @@ class CustomModule(AbstractFunction):
         self.i2c_address = None
         self.i2c_bus = None
         self.number_line_sets = None
+        self.pin_reset = None
 
         # Set custom options
         custom_function = db_retrieve_table_daemon(
@@ -316,7 +332,11 @@ class CustomModule(AbstractFunction):
             self.initialize_variables()
 
     def initialize_variables(self):
-        from mycodo.devices.lcd_generic import LCD_Generic
+        from luma.core.interface.serial import i2c
+        from luma.core.render import canvas
+        from luma.oled.device import ssd1309
+
+        self.canvas = canvas
 
         try:
             function_channels = db_retrieve_table_daemon(
@@ -331,16 +351,9 @@ class CustomModule(AbstractFunction):
 
             self.logger.debug("Line sets: {}".format(self.line_sets))
 
-            lcd_settings_dict = {
-                "unique_id": self.unique_id,
-                "i2c_address": self.i2c_address,
-                "i2c_bus": self.i2c_bus,
-                "x_characters": lcd_x_characters,
-                "y_lines": lcd_lines
-            }
-
-            self.lcd = LCD_Generic(lcd_settings_dict=lcd_settings_dict)
-            self.lcd.lcd_init()
+            self.device = ssd1309(i2c(
+                port=self.i2c_bus,
+                address=int(str(self.i2c_address), 16)))
 
             self.logger.debug("LCD Function started")
         except:
@@ -353,7 +366,7 @@ class CustomModule(AbstractFunction):
         while self.timer_loop < time.time():
             self.timer_loop += self.period
 
-        if not self.lcd:
+        if not self.device:
             self.logger.error("LCD not set up")
             return
 
@@ -367,7 +380,7 @@ class CustomModule(AbstractFunction):
             try:
                 # Get measurement value and timestamp
                 if self.options_channels['line_display_type'][current_channel] in [
-                        'measurement_value', 'measurement_ts']:
+                    'measurement_value', 'measurement_ts']:
                     lines_display[current_line] = "NONE"
                     measure_ts = None
                     measure_value = None
@@ -422,9 +435,8 @@ class CustomModule(AbstractFunction):
                     ip_out, _, _ = cmd_output(str_ip_cmd)
                     lines_display[current_line] = ip_out.rstrip().decode("utf-8")
             except Exception as err:
-                self.logger.error(
-                    "Error generating channel {} line: {}".format(
-                        current_channel, err))
+                # self.logger.exception("Generating line")
+                self.logger.error("Error generating channel {} line: {}".format(current_channel, err))
                 lines_display[current_line] = "ERROR"
 
         if self.current_line_set == len(self.line_sets) - 1:
@@ -435,11 +447,16 @@ class CustomModule(AbstractFunction):
         self.logger.debug("Displaying: {}".format(lines_display))
 
         # Display lines
-        self.lcd.lcd_init()
-        self.lcd.lcd_write_lines(
-            lines_display[0], lines_display[1], lines_display[2],lines_display[3])
+        with self.canvas(self.device) as draw:
+            draw.rectangle(self.device.bounding_box, fill="black")
+            for i, each_channel in enumerate(self.line_sets[self.current_line_set]):
+                draw.text((0, self.pad + self.line_y_dimensions[each_channel]),
+                          lines_display[each_channel], fill="white")
 
     def stop_function(self):
-        self.lcd.lcd_init()
-        self.lcd.lcd_write_lines(
-            "Mycodo {}".format(MYCODO_VERSION), "LCD Deactivated", "", "")
+        with self.canvas(self.device) as draw:
+            draw.rectangle(self.device.bounding_box, fill="black")
+            draw.text((0, self.pad + self.line_y_dimensions[0]),
+                      "Mycodo {}".format(MYCODO_VERSION), fill="white")
+            draw.text((0, self.pad + self.line_y_dimensions[1]),
+                      "Display Deactivated", fill="white")